+++ src/types/resize-ratio-calculator.ts
